Added a new ResourceConfig class and resources register into a menu.
ResourceConfig - New class to hold onto all the config and options for
a given resource.

Menu Registrations - When a resource registers itself with Active Admin
we now add it to a menu. The menus are namespaced to match the resource
namespaces. So if you create Admin::Post, a menu item will be added to
the :admin menu. If your resource is not namespaced, we default to the
:root menu