Moved initialization to ActiveAdmin.setup and removed from eager load path.
Due to the way that Rails is eager loading files, it was causing double
loading issues when not in development. Removing ourselves from the
eager load list and loading the files ourselves allows us to control
when the files get loaded as well as the number of times they are
loaded.