Rename to active_admin_master as that is what the report is for
The report isn’t for Rails 5 master but for ActiveAdmin. Seeing this
for the first time I can see someone thinking it doesn’t apply to them
because they aren’t thinking they are running into a Rails issue, but
an ActiveAdmin issue. Now we can start requiring users to use this to
replicate bugs.